feat: create, update, delete review | refactor: add sku_id when store basket

+++ Services/Catalog/YGZ.Catalog.Application/Reviews/Commands/DeleteReview/DeleteReviewHandler.cs
@@ -1,6 +1,7 @@
 ﻿using MongoDB.Driver;
 using YGZ.BuildingBlocks.Shared.Abstractions.CQRS;
 using YGZ.BuildingBlocks.Shared.Abstractions.Result;
+using YGZ.BuildingBlocks.Shared.Errors;
 using YGZ.Catalog.Application.Abstractions.Data;
 using YGZ.Catalog.Domain.Core.Errors;
 using YGZ.Catalog.Domain.Products.Iphone.Entities;
@@ -13,13 +14,13 @@ public class DeleteReviewHandler : ICommandHandler<DeleteReviewCommand, bool>
 {
     private readonly IMongoRepository<Review, ReviewId> _reviewRepository;
     private readonly OrderingProtoService.OrderingProtoServiceClient _orderingProtoServiceClient;
-    
+
     public DeleteReviewHandler(IMongoRepository<Review, ReviewId> reviewRepository, OrderingProtoService.OrderingProtoServiceClient orderingProtoServiceClient)
     {
         _reviewRepository = reviewRepository;
         _orderingProtoServiceClient = orderingProtoServiceClient;
     }
-    
+
     public async Task<Result<bool>> Handle(DeleteReviewCommand request, CancellationToken cancellationToken)
     {
 
@@ -34,20 +35,16 @@ public class DeleteReviewHandler : ICommandHandler<DeleteReviewCommand, bool>
 
         review.Delete();
 
-        // var rpcResult = await _orderingProtoServiceClient.UpdateReviewOrderItemAsync(new UpdateReviewOrderItemRquest
-        // {
-        //     ReviewId = string.IsNullOrWhiteSpace(review.Id.Value) ? null : StringValue.From(review.Id.Value),
-        //     OrderItemId = string.IsNullOrWhiteSpace(review.CustomerOrder.OrderItemId) ? null : StringValue.From(review.CustomerOrder.OrderItemId),
-        //     CustomerId = string.IsNullOrWhiteSpace(userId) ? null : StringValue.From(userId),
-        //     ReviewContent = null,
-        //     ReviewStar = null,
-        //     IsReviewed = false
-        // });
+        var rpcResult = await _orderingProtoServiceClient.UpdateOrderItemIsReviewedGrpcAsync(new UpdateOrderItemIsReviewedGrpcRequest
+        {
+            OrderItemId = review.OrderInfo.OrderItemId,
+            IsReviewed = false
+        });
 
-        //if (rpcResult.IsFailure)
-        //{
-        //    return Error.Failure(code: "OrderingGrpc.CannotUpdateReviewOrderItem", message: rpcResult.ErrorMessage?.Value ?? "Cannot update review order item", serviceName: "CatalogService");
-        //}
+        if (rpcResult.IsFailure)
+        {
+            return Result<bool>.Failure(Error.GrpcError(rpcResult.ErrorCode ?? "Unknown", rpcResult.ErrorMessage ?? "Unknown"));
+        }
 
         var result = await _reviewRepository.DeleteAsync(request.ReviewId, review, cancellationToken);
